Function calls for known functions act as annots
This means we can create a record where the fields overlap with another
record or a constructor which should be shadowed.
As an example, we can have the constructor #left in two types such as
type button and type key. A known function key-pressed? will expect a
key and recognize ctor #left as a key even if type key is shadowed by
type button.