Propagate ctor type information as annotation
This allow variants with shadowed ctors to be created when the type of
the variant is known from some source.
An example might be (type test ((#float float) (#int int) #none)) where
`#none` shadows option's `#none`. If the ctor is known to be option by
an annotation or because `#some` was matched before, `#none` refers to
the correct variant

Closes #25