Add inline attribute
We don't handle reference types properly yet. For instance a vector
might get copied by function. This creates a new vector object which
reuses the ptr of the original, but doesn't update on push.
Similarly, when getting a vector item through a function instead of
inline, a new item is created. If the object in the vector is mutable,
mutating the item returned by 'vector-get' does not return the item in
the vector. The some is true for references and the deref function.

The solution here is create special handling for reference types which
return the pointer to the underlying object (if lifetime allows it).
Until then, inlining those functions serves as a bandaid.
Implementing inlining is straight-forward enough to make this worthwhile.