Support functional record update
Syntax is `{@foo :field}` where `foo` is the name of the record to be
updated.
It's implemented in terms of normal record creation in the front end, so all existing
fields are first loaded from the old record, then stored in the new one.
No optimizing fanciness.
Surprisingly, changing the type of a polymorphic record works

Closes #15