Fix mutable parameters in tailrec functions
In tailrecursive functions, we make a fresh copy of each parameter to set
to current args in the tailrecursive call. Since we copy the values, a
mutable change will change the copy instead of the original value. This
commit fixes this by saving a pointer (casted to an int) instead of a
value of mutable structs. Instead of copying the argument to the
parameter, mutable values only set the pointer (casted to an int) to the
correct value. In the %rec block (start of the function), the ptr (int) is
casted to the corrent value, that is pointer to struct.