Group pairs into parens in match and cond
Before, in both match and cond expression the pattern / condition and
their expression were just put one after another. This looks nice for
simple cases, but makes it difficult to (automatically) format the
expression below its pattern / condition. Now, they are grouped in
parentheses like in lisp or scheme. This makes the easy case slightly
uglier, but I hope the gained benefits outweigh this.
There is also now an 'else' keyword for the last branch in an cond expression