Support mutable variables in upward closures
We split semantics between downward and upward closures.
Downward closures work as expected with reference semantics. Otherwise,
it's harder to use first-class functions.
Upward closures create another copy of the mutable state and keep it for
their own. They don't reference the mutable state from the closing
environment. I hope that's not too surprising for users. Usually,
closures are either upward or downward, for lambdas, it can't be another
way.
One can create a closure, use it in a downward way and return it
later. This case might lead to unexpected results. Let's see if it ever
comes up.
Would be best if we had a big warning (error) message for that. So far,
the handling takes place after the typing phase were errors are generated.