Restructure module name resolution
In the current implementation, [convert_var] in typing changes the
identifiers and prefixes them with a module identifier. The resulting
string is then saved and used in monomorph and throughout. The problem
with this approach is functors. In functors we want to be able to change
one module identifier (that of the functor) with another (that of the
applied functor).
To resolve this conflict the module identifier is now carried to the
monomorph pass and the string id is unchanged. In monomorph the correct
name is constructed in the [morph_var]. Event in the current
implementation name changes are needed, but only when bindings are
introduced. The new approach moves all name changes to the monomorph
pass and easily enables functors. For functors, we now have access to
the module identifier and can easily change it.