Directly use closures
The idea is the same as in the previous commit, but this time for
closures. A nice benefit here is that link order does not matter as
much. Without this commit, every module which passes (renames) a closure
needs to be linked in the correct order. This is still a TODO in our
module handling