Support def-aliases for polymorphic functions
Adds a new `Malias` case to module type. Since there is a third `Alias`
option now, a custom variant was introduced in `find_fn_callname`, which
was also renamed to `let_fn_alias`.
There was a bug in loading dependent modules as only modules which are
not already in the cache were loaded. This excludes `Located` modules.
These are now loaded correctly