Improve type inference
Set less references. Generalize function after unification with env
version, this fixes a bug found in the new hashtbl. The one link
usage (apart from normal inference) we still have is with fixed array
size. There we only link if we wouldn't create a cycle. This test is
needed with the changes to function generalization.