Fix memory leaks in pattern matches
The one memory leak was actually two issues

1. In the exclusivity pass, we only track moves for types which contain
heap allocations. The mono pass (which keeps track of the allocations
and frees them) didn't take this into account and instead tracked all
moves.
2. In the mono pass of if expressions, lexical scope wasn't respected
for binds. It's a miracle that this only lead to a bug now, this part
hasn't been touched in more than a year.