Add calls, lambdas and literals
The literals were kind of tricky. For unit `()` to be non-ambiguous, I
needed to introduce `Newline` tokens as separators for both toplevel
items as well as statement blocks. However, the syntax will be
whitespace-sensitive (using the off-side rule) anyway, thus it's the
right thing to do. And it's probably also what users expect.