Respect lexical scopes in monomorph
This is needed due to a name clash within functors, where functions and
variables can end up with the same (global in monomorph pass) name.
Luckily, we use a persistent map anyway, so it's an easy fix. In other
obvious places, this already happens. Like scopes in if, functions and lambdas.