Add `import` keyword
Instead of just finding modules like in OCaml, we have to `import` them
first. This makes it explicit what other modules the current module
uses. However, local modules of imported modules don't have to be
imported anymore. Because of this, standard library modules are just
available, because we implicitly do an `import std` and `use std` before
the current module begins.