Make move specifier optional for known functions
This is a big change. We now allow to make the move specifier `!`
optional when calling functions whose type is known. They arguments will
be treated as if they were passed by move. If the function type is not
known, e.g. in non-annotated function type parameters, it still is
needed for type inference.